Moved file classes to filesys
Classes dealing with files are now located under filesys.
This makes room for other types of "io" related classes.
